pyproject.toml
[project]
name = "sociaML"
version = "0.1.2"
description = "sociaML - the Swiss Army knife for audiovisual and textual video feature extraction."
authors = [
    {name = "David Zollikofer",email = "david@zollikofer.co"},
]
dependencies = [
    "numpy>=1.26.3",
    "pandas>=2.1.4",
    "scikit-learn>=1.3.2",
    "pyannote-audio>=3.1.1",
    "librosa>=0.10.1",
    "webvtt-py>=0.4.6",
    "spacy>=3.7.2",
    "openai-whisper>=20231117",
    "moviepy>=1.0.3",
    "tqdm>=4.66.1",
    "openai>=1.7.2",
    "seaborn>=0.13.1",
    "langchain>=0.1.0",
    "tiktoken>=0.5.2",
    "torchvision>=0.16.2",
    "nltk>=3.8.1",
    "dataclasses-json>=0.6.3",
    "sentence-transformers>=2.3.1",
    "py-feat>=0.6.1",
    "audonnx>=0.7.0",
    "audeer>=1.21.0",
    "presidio-analyzer>=2.2.351",
    "presidio-anonymizer>=2.2.351",
    "pip>=23.3.2",
    "spacy-transformers>=1.3.4",
    "transformers>=4.36.2",
    "whisperx @ git+https://github.com/m-bain/whisperx.git",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = "true"

[tool.pdm.dev-dependencies]
dev = [
    "tox-pdm>=0.7.2",
    "mypy>=1.8.0",
    "notebook>=7.0.8",
]
test = [
    "pytest>=7.4.4",
    "python-dotenv>=1.0.1",
]
